Download cover image if cache exists

+++ app/src/main/java/de/lucaspape/monstercat/download/DownloadCoverTask.kt
@@ -4,6 +4,9 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.os.AsyncTask
+import android.view.View
+import android.widget.ListView
+import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 import java.lang.IndexOutOfBoundsException
@@ -11,9 +14,10 @@ import java.lang.ref.WeakReference
 import java.net.HttpURLConnection
 import java.net.URL
 
+//TODO add imageview updater
 class DownloadCoverTask(private val weakReference: WeakReference<Context>) : AsyncTask<Void, Void, String>(){
     override fun doInBackground(vararg p0: Void?): String? {
-        val context = weakReference.get()!!
+        val context = weakReference.get()
 
         var downloadedCovers = 0
         var downloadedCoverArrays = 0
@@ -63,21 +67,23 @@ class DownloadCoverTask(private val weakReference: WeakReference<Context>) : Asy
 
     private fun downloadCover(downloadUrl: String, location: String, primaryRes:String, secondaryRes:String):Boolean{
         try {
-            val url = URL("$downloadUrl?image_width=$primaryRes")
-            val connection = url.openConnection() as HttpURLConnection
-            connection.doInput = true
-            connection.connect()
-            val input = connection.inputStream
-            val primaryBitmap = BitmapFactory.decodeStream(input)
+            if(!File(location + primaryRes).exists() || !File(location + secondaryRes).exists()){
+                val url = URL("$downloadUrl?image_width=$primaryRes")
+                val connection = url.openConnection() as HttpURLConnection
+                connection.doInput = true
+                connection.connect()
+                val input = connection.inputStream
+                val primaryBitmap = BitmapFactory.decodeStream(input)
 
-            FileOutputStream(location + primaryRes).use { out ->
-                primaryBitmap!!.compress(Bitmap.CompressFormat.PNG, 100, out)
-            }
+                FileOutputStream(location + primaryRes).use { out ->
+                    primaryBitmap!!.compress(Bitmap.CompressFormat.PNG, 100, out)
+                }
 
-            val secondaryBitmap = Bitmap.createScaledBitmap(primaryBitmap, secondaryRes.toInt(), secondaryRes.toInt(), false)
+                val secondaryBitmap = Bitmap.createScaledBitmap(primaryBitmap, secondaryRes.toInt(), secondaryRes.toInt(), false)
 
-            FileOutputStream(location + secondaryRes).use { out ->
-                secondaryBitmap!!.compress(Bitmap.CompressFormat.PNG, 100, out)
+                FileOutputStream(location + secondaryRes).use { out ->
+                    secondaryBitmap!!.compress(Bitmap.CompressFormat.PNG, 100, out)
+                }
             }
 
             return true
